Extract askama_parser from askama_derive
This is a rote split of the proc_macro crate into the config and
parser side on the one hand, and the generator code on the other.
