+++ Cargo.toml
@@ -3,10 +3,11 @@ members = [
     "askama",
     "askama_actix",
     "askama_axum",
-    "askama_gotham",
     "askama_derive",
     "askama_escape",
+    "askama_gotham",
     "askama_mendes",
+    "askama_parser",
     "askama_rocket",
     "askama_tide",
     "askama_warp",
