+++ askama_parser/Cargo.toml
@@ -1,16 +1,13 @@
 [package]
-name = "askama_derive"
-version = "0.12.0"
-description = "Procedural macro package for Askama"
+name = "askama_parser"
+version = "0.1.0"
+description = "Askama template syntax parser"
 homepage = "https://github.com/djc/askama"
 repository = "https://github.com/djc/askama"
 license = "MIT/Apache-2.0"
 workspace = ".."
 readme = "README.md"
 edition = "2018"
-
-[lib]
-proc-macro = true
 
 [features]
 config = ["serde", "toml"]
@@ -30,7 +27,6 @@ with-tide = []
 with-warp = []
 
 [dependencies]
-askama_parser = { version = "0.1.0", path = "../askama_parser" }
 mime = "0.3"
 mime_guess = "2"
 nom = "7"
