+++ askama_parser/LICENSE-APACHE
@@ -1,0 +1,1 @@
+../LICENSE-APACHE
