+++ askama_parser/LICENSE-MIT
@@ -1,0 +1,1 @@
+../LICENSE-MIT
