+++ askama_parser/README.md
@@ -1,0 +1,9 @@
+# askama_derive: procedural macros for the Askama templating engine
+
+[![Documentation](https://docs.rs/askama_derive/badge.svg)](https://docs.rs/askama_derive/)
+[![Latest version](https://img.shields.io/crates/v/askama_derive.svg)](https://crates.io/crates/askama_derive)
+[![Build Status](https://github.com/djc/askama/workflows/CI/badge.svg)](https://github.com/djc/askama/actions?query=workflow%3ACI)
+[![Chat](https://badges.gitter.im/gitterHQ/gitter.svg)](https://gitter.im/djc/askama)
+
+This crate contains the procedural macros used by the
+[Askama](https://github.com/djc/askama) templating engine.
