+++ askama_parser/templates/a.html
@@ -1,0 +1,1 @@
+foo
