+++ askama_parser/templates/b.html
@@ -1,0 +1,1 @@
+bar
