+++ askama_parser/templates/sub/b.html
@@ -1,0 +1,1 @@
+bar
